LayoutTests/imported/w3c/web-platform-tests/referrer-policy/4K+1/gen/top.meta/strict-origin-when-cross-origin/a-tag.http-expected.txt



FAIL Referrer Policy: Expects stripped-referrer for a-tag to same-http origin and no-redirect redirection from http context. assert_equals: Reported Referrer URL is 'origin'. expected "http://localhost:8800/" but got "http://localhost:8800"
